requirements.txt
astroid==1.6.0
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dill==0.2.7.1
dj-database-url==0.4.2
Django==2.0.1
django-extensions==1.9.9
django-filter==1.1.0
django-heroku==0.2.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
firebase-admin==2.7.0
future==0.16.0
grpcio==1.8.4
gunicorn==19.7.1
idna==2.6
isort==4.2.15
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==2.6.10
MarkupSafe==1.0
mccabe==0.6.1
openapi-codec==1.3.2
ply==3.8
protobuf==3.5.1
psycopg2==2.7.3.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pylint==1.8.1
pytz==2017.3
requests==2.18.4
rsa==3.4.2
simplejson==3.13.2
six==1.11.0
typing==3.6.2
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
wrapt==1.10.11
yapf==0.20.0
django-cloudinary-storage==0.2.3
cloudinary==1.8.0
Pillow==5.0.0
django-phonenumber-field==2.0.0